+++ Makefile
@@ -7,4 +7,6 @@ wrap:
 	docker compose run transformer python scripts/wrap_tables.py
 
 dbt: build
+	docker compose run transformer dbt run --target optimism_mainnet --profiles-dir profiles --profile docker
 	docker compose run transformer dbt run --target base_goerli --profiles-dir profiles --profile docker
+	docker compose run transformer dbt run --target base_mainnet --profiles-dir profiles --profile docker
