+++ dashboard/.dockerignore
@@ -1,0 +1,3 @@
+.env
+/env
+__pycache__/
