+++ dashboard/Home.py
@@ -1,8 +1,6 @@
 import streamlit as st
 
-st.set_page_config(
-    page_title="Synthetix Dashboards",
-)
+st.set_page_config(page_title="Synthetix Dashboards", layout="wide")
 
 hide_footer = """
     <style>
@@ -15,6 +13,6 @@ st.write("# Synthetix Dashboards")
 
 st.markdown(
     """
-Use the sidebar to select a dashboard.
+Use the sidebar to select a chain. The dashboard selector will appear below.
 """
 )
