+++ dashboard/utils/charts.py
@@ -4,17 +4,32 @@ import plotly.express as px
 ## charts
 def chart_bars(df, x_col, y_cols, title, color=None):
     fig = px.bar(df, x=x_col, y=y_cols, title=title, color=color)
+    fig.update_traces(hovertemplate=None)
+    fig.update_layout(
+        hovermode="x unified",
+        legend=dict(orientation="h", yanchor="bottom", xanchor="left", y=-0.4),
+    )
     return fig
 
 
 def chart_lines(df, x_col, y_cols, title, color=None):
     fig = px.line(df, x=x_col, y=y_cols, title=title, color=color)
+    fig.update_traces(hovertemplate=None)
+    fig.update_layout(
+        hovermode="x unified",
+        legend=dict(orientation="h", yanchor="bottom", xanchor="left", y=-0.4),
+    )
     return fig
 
 
 # with asset filter
 def chart_asset_bars(df, asset, x_col, y_cols, title):
     fig = px.bar(df[df["asset"] == asset], x=x_col, y=y_cols, title=f"{title}: {asset}")
+    fig.update_traces(hovertemplate=None)
+    fig.update_layout(
+        hovermode="x unified",
+        legend=dict(orientation="h", yanchor="bottom", xanchor="left", y=-0.4),
+    )
     return fig
 
 
@@ -25,6 +40,11 @@ def chart_asset_lines(df, asset, x_col, y_cols, title):
         y=y_cols,
         line_shape="hv",
         title=f"{title}: {asset}",
+    )
+    fig.update_traces(hovertemplate=None)
+    fig.update_layout(
+        hovermode="x unified",
+        legend=dict(orientation="h", yanchor="bottom", xanchor="left", y=-0.4),
     )
     return fig
 
@@ -38,6 +58,11 @@ def chart_asset_oi(df, asset):
         color_discrete_sequence=["red", "green"],
         title=f"Open Interest %: {asset}",
     )
+    fig.update_traces(hovertemplate=None)
+    fig.update_layout(
+        hovermode="x unified",
+        legend=dict(orientation="h", yanchor="bottom", xanchor="left", y=-0.4),
+    )
     return fig
 
 
@@ -45,6 +70,11 @@ def chart_asset_oi(df, asset):
 def chart_market_bars(df, asset, x_col, y_cols, title):
     fig = px.bar(
         df[df["market_symbol"] == asset], x=x_col, y=y_cols, title=f"{title}: {asset}"
+    )
+    fig.update_traces(hovertemplate=None)
+    fig.update_layout(
+        hovermode="x unified",
+        legend=dict(orientation="h", yanchor="bottom", xanchor="left", y=-0.4),
     )
     return fig
 
@@ -57,6 +87,11 @@ def chart_market_lines(df, asset, x_col, y_cols, title):
         line_shape="hv",
         title=f"{title}: {asset}",
     )
+    fig.update_traces(hovertemplate=None)
+    fig.update_layout(
+        hovermode="x unified",
+        legend=dict(orientation="h", yanchor="bottom", xanchor="left", y=-0.4),
+    )
     return fig
 
 
@@ -68,5 +103,10 @@ def chart_market_oi(df, asset):
         line_shape="hv",
         color_discrete_sequence=["red", "green"],
         title=f"Open Interest %: {asset}",
+    )
+    fig.update_traces(hovertemplate=None)
+    fig.update_layout(
+        hovermode="x unified",
+        legend=dict(orientation="h", yanchor="bottom", xanchor="left", y=-0.4),
     )
     return fig
